updating tea code to run with latest correctionlib

+++ libs/extensions/src/NanoEventProcessor.cpp
@@ -222,8 +222,8 @@ map<string, float> NanoEventProcessor::GetDSAMuonEfficiencyScaleFactors(const sh
   auto& scaleFactorsManager = ScaleFactorsManager::GetInstance();
   bool firstIteration = true;
   for (auto muon : *muonCollection) {
-    vector<variant<int, double, string>> args = {double(muon->Get("pt"))};
-    // vector<variant<int, double, string>> args = {double(muon->Get("pt")), double(fabs(muon->GetAs<float>("dxyPVTraj")))};
+    vector<variant<long, double, string>> args = {double(muon->Get("pt"))};
+    // vector<variant<long, double, string>> args = {double(muon->Get("pt")), double(fabs(muon->GetAs<float>("dxyPVTraj")))};
     if (firstIteration) {
       auto weights_setup = scaleFactorsManager.GetCustomScaleFactors("DSAEff", args);
       for (auto& [name, weight] : weights_setup) weights[name] = 1.0;
