+++ libs/extensions/src/NanoMuon.cpp
@@ -75,11 +75,11 @@ map<string,float> NanoMuon::GetScaleFactors(string nameID, string nameIso, strin
 map<string,float> NanoMuon::GetEmptyDSAScaleFactors(string nameID, string nameReco_cosmic) {
   auto &scaleFactorsManager = ScaleFactorsManager::GetInstance();
 
-  vector<variant<int, double, string>> args_jpsi = {fabs(GetEta()), GetPt()};
+  vector<variant<long, double, string>> args_jpsi = {fabs(GetEta()), GetPt()};
   string nameID_jpsi = nameID;
   map<string,float> idSF_jpsi = scaleFactorsManager.GetDSAMuonScaleFactors(nameID_jpsi, args_jpsi);
 
-  vector<variant<int, double, string>> args_reco = {};
+  vector<variant<long, double, string>> args_reco = {};
   map<string,float> recoSF = scaleFactorsManager.GetDSAMuonScaleFactors(nameReco_cosmic, args_reco);
 
   map<string,float> emptySF;
@@ -97,11 +97,11 @@ map<string,float> NanoMuon::GetDSAScaleFactors(string nameID, string nameReco_co
 
   auto &scaleFactorsManager = ScaleFactorsManager::GetInstance();
 
-  vector<variant<int, double, string>> args_jpsi = {fabs(GetEta()), GetPt()};
+  vector<variant<long, double, string>> args_jpsi = {fabs(GetEta()), GetPt()};
   string nameID_jpsi = nameID;
   map<string,float> idSF_jpsi = scaleFactorsManager.GetDSAMuonScaleFactors(nameID_jpsi, args_jpsi);
 
-  vector<variant<int, double, string>> args_cosmic = {fabs((float)Get("dxyPVTraj"))};
+  vector<variant<long, double, string>> args_cosmic = {fabs((float)Get("dxyPVTraj"))};
   string nameID_cosmic = nameID + "_cosmic";
   map<string,float> idSF_cosmic = scaleFactorsManager.GetDSAMuonScaleFactors(nameID_cosmic, args_cosmic);
 
@@ -113,7 +113,7 @@ map<string,float> NanoMuon::GetDSAScaleFactors(string nameID, string nameReco_co
     string name_cosmic = nameID_cosmic + "_" + variation;
     idSF[name_jpsi] = idSF_jpsi[name_jpsi] * idSF_cosmic[name_cosmic];
   }
-  vector<variant<int, double, string>> args_reco = {};
+  vector<variant<long, double, string>> args_reco = {};
   map<string,float> recoSF = scaleFactorsManager.GetDSAMuonScaleFactors(nameReco_cosmic, args_reco);
 
   scaleFactor["systematic"] = idSF["systematic"] * recoSF["systematic"];
